reorder test and fix asserts
