+++ tests/operators/test_constraints.py
@@ -16,18 +16,7 @@
 
 from datetime import datetime
 
-import iris
-import iris.cube
-import pytest
-
 from CSET.operators import constraints
-
-
-# Session scope fixtures, so the test data only has to be loaded once.
-@pytest.fixture(scope="session")
-def load_verticalcoord_cubelist() -> iris.cube.CubeList:
-    """Get a cubelist with multiple vertical level cubes."""
-    return iris.load("tests/test_data/vertlevtestdata.nc", "y_wind")
 
 
 def test_generate_stash_constraint():
@@ -95,52 +84,13 @@ def test_generate_level_constraint_multi_level():
     assert expected_pressure_constraint in repr(pressure_constraint)
 
 
-def test_generate_level_constraint_returnsinglelevel(load_verticalcoord_cubelist):
-    """For a cubelist that contains 3 cubes on different vertical levels.
-
-    Return one without a vertical coordinate.
-    """
-    constraint_1 = constraints.generate_level_constraint(
-        coordinate="pressure", levels=[]
-    )
-    constraint_2 = constraints.generate_level_constraint(
-        coordinate="model_level_number", levels=[]
-    )
-    combined = constraints.combine_constraints(constraint_1, a=constraint_2)
-
-    extracted = load_verticalcoord_cubelist.extract(combined)[0]
-
-    assert str(
-        extracted.coord
-    ), "<bound method Cube.coords of <iris 'Cube' of y_wind / (m s-1) (latitude: 2; longitude: 2)>>"
-
-
-def test_generate_level_constraint_returnallpressure(load_verticalcoord_cubelist):
-    """For a cubelist that contains 3 cubes on different vertical levels.
-
-    Return one with a pressure on all levels.
-    """
-    constraint_1 = constraints.generate_level_constraint(
+def test_generate_level_constraint_all_level():
+    """Generate constraint for all levels."""
+    pressure_constraint = constraints.generate_level_constraint(
         coordinate="pressure", levels="*"
     )
-
-    assert str(
-        load_verticalcoord_cubelist.extract(constraint_1)[0].coord
-    ), "<bound method Cube.coord of <iris 'Cube' of y_wind / (m s-1) (pressure: 34; latitude: 2; longitude: 2)>>"
-
-
-def test_generate_level_constraint_returnallmodlev(load_verticalcoord_cubelist):
-    """For a cubelist that contains 3 cubes on different vertical levels.
-
-    Return one with a model level on all levels.
-    """
-    constraint_1 = constraints.generate_level_constraint(
-        coordinate="model_level_number", levels="*"
-    )
-
-    assert str(
-        load_verticalcoord_cubelist.extract(constraint_1)[0].coord
-    ), "<bound method Cube.coord of <iris 'Cube' of y_wind / (m s-1) (model_level_number: 70; latitude: 2; longitude: 2)>>"
+    expected_pressure_constraint = "Constraint(coord_values={'pressure': <function generate_level_constraint.<locals>.<lambda> at"
+    assert expected_pressure_constraint in repr(pressure_constraint)
 
 
 def test_generate_level_constraint_no_pressure():
